refactor[#345] : 펫정보 변경 api 요청 함수에 apiClient 적용
