+++ src/features/setting/api/index.ts
@@ -5,4 +5,4 @@ export * from "./putChangePassword";
 export * from "./putSetPassword";
 export * from "./putChangeAge";
 export * from "./putChangeGender";
-export * from "./putChangePetInformation";
+export * from "./putChangePetInfo";
