+++ src/features/setting/constants/endpoint.ts
@@ -8,5 +8,5 @@ export const SETTING_END_POINT = {
   CHANGE_PASSWORD: "/users/profile/password",
   SET_PASSWORD: `${API_BASE_URL}/users/profile/password/social`,
   CHANGE_AGE: "/users/profile/age",
-  CHANGE_PET_INFO: `${API_BASE_URL}/pets`,
+  CHANGE_PET_INFO: "/pets",
 };
